refactor: PLAN フェーズ A/B 一括対応
A-3: ToPaneTarget/ToPaneZone helper を app_events.h に追加し reducer の if/else を集約
A-4: DrawSidePaneImpl の 17 引数を SidePaneDrawContext 構造体に集約
A-5: WebView2 lib path を x64/arm64 で分岐 (それ以外は FATAL_ERROR)
B-1: SetTimer/KillTimer/FileWatcher 直呼びを EmitEffect 経路に統一。EmitEffect は ExecuteOne 直送に変更し pmr vector の per-call alloc を排除
B-2: OnParseComplete/DoReloadCurrentFile の reload diff 早期 return を ApplyReloadDecisionEarly helper に集約
B-3: document_utils.cpp を selection_html / reload_scroll / reload_diff の 3 ファイルへ分割
B-4: FindAnchorNodeIndex / TableOfContents::BuildFromNodes をテスト用 document_test_helpers.h に切り出し
A-1: DWrite テストを DWriteTestBase 共通基底に統一

/simplify レビュー反映:
- Document::BuildHeadingIndices に toc_.Reserve() を追加し旧 BuildFromNodes の reserve heuristic を復元
- anchor_search.{h,cpp} を reload_scroll.{h,cpp} に rename (FindAnchorNodeIndex 移動後の実態に合わせ reload_diff の補助という位置づけを明確化)
