+++ src/app/app_events.h
@@ -8,6 +8,7 @@
 #include <string>
 #include <cstdint>
 #include <memory_resource>
+#include <optional>
 
 // プラットフォーム非依存のピクセル矩形。App::OnDpiChanged 境界で
 // Win32 の RECT から各フィールドの値をコピーして生成する。
@@ -73,6 +74,23 @@ enum class PaneTarget {
 struct TogglePaneAction {
     PaneTarget target;
 };
+
+// PaneZone (座標ヒット判定の結果) を、サイドペイン操作の対象を表す
+// PaneTarget へ変換する。File/Toc 以外の領域 (None / Splitter / MdPane)
+// は対象外なので nullopt を返す。
+constexpr std::optional<PaneTarget> ToPaneTarget(PaneZone zone) noexcept
+{
+    switch (zone) {
+    case PaneZone::FilePane: return PaneTarget::File;
+    case PaneZone::TocPane:  return PaneTarget::Toc;
+    default:                 return std::nullopt;
+    }
+}
+
+constexpr PaneZone ToPaneZone(PaneTarget target) noexcept
+{
+    return target == PaneTarget::File ? PaneZone::FilePane : PaneZone::TocPane;
+}
 
 // ズーム操作
 enum class ZoomDirection {
